move out rand; tiny refactor for WASM; alpha release
