+++ calcit/snapshots/test-list.cirru
@@ -124,12 +124,6 @@
                   fn (x) (&> x 3)
                 [] 0 1 2 3
 
-              assert-detect identity $ <= 0
-                index-of (range 10) $ rand-nth $ range 10
-
-              assert= nil $ rand-nth ([])
-              assert= nil (;nil anything)
-
               assert-detect identity $ contains? (range 10) 6
               assert-detect not $ contains? (range 10) 16
